requirements.txt
openai>=1.1.0
mcp>=0.1.0
python-dotenv>=1.0.0
asyncio>=3.4.3
